acre: Acoustic spatial capture-recapture models

The acre package provides functions to fit spatial capture-recapture models, with a particular focus on analysing data from acoustic surveys.

See also

Author

Maintainer: Ben Stevenson ben.stevenson@auckland.ac.nz

Authors:

Ben Stevenson ben.stevenson@auckland.ac.nz

Joseph Reps jrep150@aucklanduni.ac.nz

Lingyu Hao lhao446@aucklanduni.ac.nz