Title

Title

Usage

# S3 method for class 'acreboot'
confint(
  object,
  parm = NULL,
  level = 0.95,
  types = NULL,
  new.covariates = NULL,
  correct_bias = FALSE,
  from_boot = TRUE,
  ...
)

Arguments

object

an object generated by the bootstrap function "boot.acre()".

parm

a character vector, the same as "coef.acre()".

level

a numeric value indicates the confident level, default is 0.95.

types

a character vector, the same as "coef.acre()".

new.covariates

a data frame, the same as "coef.acre()".

correct_bias

a logical value indicates whether to apply a bias correction method to the confidence interval. Default is FALSE, and the naive percentiles from the bootstrap results will used as confidence interval; if TRUE, the bootstrap results will be corrected by beta_boot = 2 * beta_est - beta_boot, where beta_boot is the bootstrap results values and beta_est is the estimations from "coef.acre()", and the percentiles from these corrected bootstrap results are used as the confidence interval.

from_boot

a logical value indicates whether to use the bootstrap results to construct the confidence matrix. Default is TRUE; if FALSE, "confint.acre()" will be called.