Fitting a spatial capture-recapture model

Fits an a spatial capture-recapture model to a data object returned by read.acre().

Usage

fit.acre(
  data,
  model = NULL,
  detfn = NULL,
  sv = NULL,
  bounds = NULL,
  fix = NULL,
  ss.opts = NULL,
  control.mask = NULL,
  mask = NULL,
  convert.loc2mask = list(),
  scale.covs = TRUE,
  local = FALSE,
  tracing = TRUE,
  gr.skip = FALSE,
  two.stage = FALSE,
  CL = two.stage
)

Arguments

data

A data object returned by read.acre().

model

A list with named components. Each component name must match a parameter name. The component itself must be a formula specifying the relationship between covariates and the parameter. See the section on model specifications below.

detfn

A character string specifying the detection function to be used. Either "hn" (halfnormal), "hhn" (hazard halfnormal), "hr" (hazard rate), "th" (threshold), "lth" (log-link threshold), or "ss" (signal strength). If "ss" is used, signal strength information must be included in data. See the section below on parameter names for further details.

sv

A named list of parameter start values. Element names are parameter names, and each element is a start value for the associated parameter. See the section below on troubleshooting model fitting for further details.

bounds

A named list of parameter bounds. Element names are parameter names, and each element is a vector of length two, specifying the upper and lower bounds for the associated parameter. See the section below on troubleshooting model fitting for further details.

fix

A named list of fixed parameter values. Element names are parameter names to be fixed, and each element is the fixed value for the associated parameter. If a parameter is fixed, it will not be estimated by the model.

ss.opts

A list with information required to fit models that include signal strengths as auxiliary detection data. One component must be named cutoff, a detection threshold. It could contain 3 elements related to signal strength model. An optional component is ss.link, which specifies the relationship between distance and the expected received signal strength. See the section below on signal strength models for further details.

control.mask

a list with elements corresponding to the function create.mask. If provided, and the argument "mask" is not provided, it will be used together with "traps" in the "dat" to expand masks.

mask

an object generated by the function create.mask(), if not provided, this function will create it based on provided traps and control.mask.

convert.loc2mask

a list. When either the arguments of mask or control.mask is provided, the mask will be created by them, and if there is any parameter extended under mask level, this argument could be used to control the process of converting location related covariates to the new mask level data. For any details, could refer to the help document in the function read.acre()

scale.covs

a logical value. Indicate whether to standardize the numerical covariates for the extended parameters, it is TRUE by default.

local

a logical value. FALSE by default. If TRUE, the model will only integrate the masks within the buffer distance for all traps.

tracing

a logical value. TRUE by default, an indicator of showing the tracing information. (seems not work as expected.)

gr.skip

a logical value. FALSE by default. If TRUE, TMB model will skip the process of generating automatic derivative functions, which will use less RAM, but the optimization process will consume more time.

two.stage

a logical value. FALSE by default. If TRUE, after fitting a model, the partial derivatives, of the ESA for each session, w.r.t. the other parameters, is calculated and returned.

CL

a logical value. FALSE by default. If TRUE, fit a conditional likelihood model, ignoring density parameter estimation.

Model specification

Specify models using the model argument.

Parameter names

Parameters fall into three groups:

D, density (in calls or animals per hectare).

Detection function parameters.

Auxiliary data measurement error parameters.

(More to put in here)

Troubleshooting model fiting

Sometimes model-fitting might be tricky. Setting start values or applying bounds on parameters might help.

(More to put in here)

Signal strength models

There are a few options available for signal strength models.

(More to put in here)