Original fit function from ascr

Original fit function from ascr

Usage

fit_og(
  capt,
  traps,
  mask,
  detfn = NULL,
  sv = NULL,
  bounds = NULL,
  fix = NULL,
  ss.opts = NULL,
  cue.rates = NULL,
  survey.length = NULL,
  sound.speed = 331,
  local = FALSE,
  par.extend = NULL,
  tracing = TRUE,
  gr.skip = FALSE,
  sv.link = NULL,
  CL = FALSE,
  two.stage = FALSE,
  ...
)

Arguments

capt

a list or a data frame with capture data. Columns named session, ID and trap are required, with each row being regarded as one detection. Extra information can be provided as columns bearing, dist ,ss, toa. animal_ID could be provided if individuals could be distinguished from their acoustic detection. In addition, any individuals from different sessions will be treated as independent ones. The default unit for any distance related value should be 'meter', and for time related value should be 'seconds'..

traps

a data frame or a list, the same as the argument 'trap' in function create.capt(), please refer to the description for that function.

mask

a list or a data frame. Generated by the function create.mask().

detfn

same as the argument with the same name in the function read.acre().

sv

same as the argument with the same name in the function read.acre().

bounds

same as the argument with the same name in the function read.acre().

fix

same as the argument with the same name in the function read.acre().

ss.opts

same as the argument with the same name in the function read.acre().

cue.rates

same as the argument with the same name in the function read.acre().

survey.length

same as the argument with the same name in the function read.acre().

sound.speed

same as the argument with the same name in the function read.acre().

local

same as the argument with the same name in the function fit.acre().

par.extend

a list. contains elements of model, data, scale and link. model: the same as argument model in the function fit.acre(); data: a list, could contain elements of session, trap, mask. Each of them should be a data frame contains the corresponding information of sessions, traps or masks and the corresponding covariates; scale: the same as argument is.scale in the function fit.acre(); link: the same as argument model.link in the function fit.acre().

tracing

same as the argument with the same name in the function fit.acre().

gr.skip

same as the argument with the same name in the function fit.acre().

sv.link

a list; this is mostly for development purpose, not recommended to use.

CL

same as the argument with the same name in the function fit.acre().

two.stage

same as the argument with the same name in the function fit.acre().