Fitting acoustic SCR models

Fits an acoustic SCR model.

Usage

fit.acre(
  dat,
  model = NULL,
  detfn = NULL,
  sv = NULL,
  bounds = NULL,
  fix = NULL,
  ss.opts = NULL,
  control.mask = NULL,
  mask = NULL,
  convert.loc2mask = list(),
  is.scale = TRUE,
  model.link = NULL,
  local = FALSE,
  tracing = TRUE,
  gr.skip = FALSE,
  sv.link = NULL,
  CL = FALSE
)

Arguments

dat

an object generated by the function read.acre()

model

a list with named elements. Each element contains the formula for the parameter indicated by the name of this element. For example, model = list(D = ~x1+x2, sigma = ~x3+x4)

detfn

character string specifying the detection function to be used. One of "hn(halfnormal),hhn(hazard halfnormal),hr(hazard rate),th(threshold),lth(log-link threshold), orss(signal strength). Ifss` is used, signal strength information must be provided in capt.

sv

a named list of parameter start values. Element names are parameter names, and each element is a start value for the associated parameter. See 'Details' for further information on the parameters to be fitted.

bounds

A named list of parameter bounds. Element names are parameter names, and each element is a vector of length two, specifying the upper and lower bounds for the associated parameter.

fix

a named list of fixed parameter values. Element names are parameter names to be fixed, and each element is the fixed value for the associated parameter. If a parameter is fixed, it will not be estimated by the model.

ss.opts

a list. It could contain 3 elements related to signal strength model. cutoff - the threshold of signal strength could be detected, must be included if signal strength is provided. ss.link - the link function of signal strength, could be identity, log or spherical, by default, it is identity. directional - a logical value, indicate whether to apply signal strength directional model.

control.mask

a list with elements corresponding to the function create.mask. If provided, and the argument "mask" is not provided, it will be used together with "traps" in the "dat" to expand masks.

mask

an object generated by the function create.mask(), if not provided, this function will create it based on provided traps and control.mask.

convert.loc2mask

a list. When either the arguments of mask or control.mask is provided, the mask will be created by them, and if there is any parameter extended under mask level, this argument could be used to control the process of converting location related covariates to the new mask level data. For any details, could refer to the help document in the function read.acre()

is.scale

a logical value. Indicate whether to standardize the numerical covariates for the extended parameters, it is TRUE by default.

model.link

a list with named elements. Each element contains the link for the parameter indicated by the name of this element.

local

a logical value. FALSE by default. If TRUE, the model will only integrate the masks within the buffer distance for all traps.

tracing

a logical value. TRUE by default, an indicator of showing the tracing information. (seems not work as expected.)

gr.skip

a logical value. FALSE by default. If TRUE, TMB model will skip the process of generating automatic derivative functions, which will use less RAM, but the optimization process will consume more time.

sv.link

a list; this is mostly for development purpose, not recommended to use

CL

a logical value. FALSE by default. If TRUE, fit a conditional likelihood model, ignoring density parameter estimation.